requirements.txt
django-allauth==0.29.0
djangorestframework
apiclient==1.0.3
Django==1.10.6
google-api-python-client==1.6.2
httplib2==0.10.3
mock==2.0.0
oauth2client==4.0.0
oauthlib==2.0.1
python-openid==2.2.5
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
urllib3==1.20
